Move cluster controller proto to core
The benefit being that you don't have to import the admin crate, which
is pretty heavy, just to get a cluster controller client.
