+++ crates/admin/Cargo.toml
@@ -70,9 +70,6 @@ tower = { workspace = true, features = ["load-shed", "limit"] }
 tracing = { workspace = true }
 xxhash-rust = { workspace = true }
 
-[build-dependencies]
-tonic-build = { workspace = true }
-
 [dev-dependencies]
 restate-bifrost = { workspace = true, features = ["test-util"] }
 restate-core = { workspace = true, features = ["test-util"] }
