+++ crates/core/build.rs
@@ -16,6 +16,18 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     tonic_build::configure()
         .bytes(["."])
+        .file_descriptor_set_path(out_dir.join("cluster_ctrl_svc_descriptor.bin"))
+        // allow older protobuf compiler to be used
+        .protoc_arg("--experimental_allow_proto3_optional")
+        .extern_path(".restate.common", "::restate_types::protobuf::common")
+        .extern_path(".restate.cluster", "::restate_types::protobuf::cluster")
+        .compile_protos(
+            &["./protobuf/cluster_ctrl_svc.proto"],
+            &["protobuf", "../types/protobuf"],
+        )?;
+
+    tonic_build::configure()
+        .bytes(["."])
         .file_descriptor_set_path(out_dir.join("node_ctl_svc_descriptor.bin"))
         // allow older protobuf compiler to be used
         .protoc_arg("--experimental_allow_proto3_optional")
