+++ crates/core/protobuf/cluster_ctrl_svc.proto
