+++ server/Cargo.toml
@@ -75,7 +75,7 @@ url = { version = "2.5.4", features = [] }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 [dev-dependencies]
-restate-admin = { workspace = true, features = ["memory-loglet", "clients"] }
+restate-admin = { workspace = true, features = ["memory-loglet"] }
 restate-bifrost = { workspace = true, features = ["test-util"] }
 restate-core = { workspace = true, features = ["test-util"] }
 restate-local-cluster-runner = { workspace = true }
