+++ tools/restatectl/src/commands/log/trim_log.rs
@@ -12,8 +12,8 @@ use anyhow::Context;
 use cling::prelude::*;
 use tracing::error;
 
-use restate_admin::cluster_controller::protobuf::{TrimLogRequest, new_cluster_ctrl_client};
 use restate_cli_util::c_println;
+use restate_core::protobuf::cluster_ctrl_svc::{TrimLogRequest, new_cluster_ctrl_client};
 use restate_types::nodes_config::Role;
 
 use crate::connection::ConnectionInfo;
