+++ tools/restatectl/src/commands/status.rs
@@ -18,10 +18,10 @@ use itertools::Itertools;
 use tonic::{Code, IntoRequest};
 use tracing::{error, warn};
 
-use restate_admin::cluster_controller::protobuf::{ClusterStateRequest, new_cluster_ctrl_client};
 use restate_cli_util::_comfy_table::{Cell, Color, Row, Table};
 use restate_cli_util::c_println;
 use restate_cli_util::ui::console::StyledTable;
+use restate_core::protobuf::cluster_ctrl_svc::{ClusterStateRequest, new_cluster_ctrl_client};
 use restate_metadata_server::grpc::new_metadata_server_client;
 use restate_types::health::MetadataServerStatus;
 use restate_types::logs::metadata::Logs;
